pyproject.toml
[project]
name = "hexanavigator"
version = "0.1.0"
description = "Hexapod autonomous navigation scaffold for Raspberry Pi with QGC mission following"
authors = [
    { name = "CyberCastle", email = "devnull@example.com" }
]
license = { text = "LGPL" }
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "geographiclib>=2.0,<3.0",
    "numpy>=2.0,<3.0",
    "pydantic>=2.8,<3.0",
    "pyyaml>=6.0.1,<7.0.0",
    "requests>=2.32,<3.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.8.0,<25.0.0",
    "pytest>=8.3.3,<9.0.0",
    "ruff>=0.6.9,<1.0.0",
    "mypy>=1.11.2,<2.0.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
# Manage dependencies with Poetry without building this package by default
package-mode = false

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
lint.select = ["E", "F", "I", "UP", "D"]
lint.ignore = ["D1"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
strict_optional = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true